Broke down the Executor into modules and protocols
The code is now in smaller chunks and easier to understand but tests now
run 2 seconds slower on my MacBook Pro. (4.0s -> 6.0s).

I think this is mostly down to the introduction Completion protocol
which replaced all of the complete_value methods in the Executor.
