+++ lib/graphql/type/non_null.ex
@@ -7,3 +7,12 @@ defmodule GraphQL.Type.NonNull do
   end
 end
 
+defimpl GraphQL.Execution.Completion, for: GraphQL.Type.NonNull do
+  alias GraphQL.Execution.Completion
+  alias GraphQL.Execution.Types
+
+  @spec complete_value(%GraphQL.Type.NonNull{}, ExecutionContext.t, GraphQL.Document.t, any, any) :: {ExecutionContext.t, map}
+  def complete_value(%GraphQL.Type.NonNull{ofType: inner_type}, context, field_asts, info, result) do
+    Completion.complete_value(Types.unwrap_type(inner_type), context, field_asts, info, result)
+  end
+end
