+++ lib/graphql/type/object.ex
@@ -11,4 +11,13 @@ defmodule GraphQL.Type.ObjectType do
   defimpl String.Chars do
     def to_string(obj), do: obj.name
   end
+
+  defimpl GraphQL.Execution.Completion do
+    alias GraphQL.Execution.Selection
+
+    def complete_value(return_type, context, field_asts, _info, result) do
+      Selection.complete_sub_fields(return_type, context, field_asts, result)
+    end
+  end
 end
+
